Remove unused routes, Add create-account route and web page
Having forked the repo, it became apparent that a lot of the code was not going to be useful for the coding challenge. I removed some routes that were unnecessary, created a create-account route and react component, as well as
made it to that if you were going to the root of the web page, you'd be re-routed to the create-account page.